Refactor index.ts by splitting logic into src/handlers/
